:pencil2: move `register_windclass()` to `struct Window`
