+++ src/main.rs
@@ -24,9 +24,6 @@ fn main() {
     unsafe {
         let class_name = encode("my_window_class_name");
         let window_name = encode("Hello, World!");
-        if !register_wndclass(&class_name) {
-            return;
-        }
 
         let window = match Window::create(&class_name, &window_name) {
             Ok(result) => result,
@@ -50,29 +47,4 @@ fn main() {
 
 fn encode(source: &str) -> Vec<u16> {
     source.encode_utf16().chain(Some(0)).collect()
-}
-
-unsafe fn register_wndclass(class_name: &[u16]) -> bool {
-    let mut winc = mem::zeroed::<WNDCLASSW>();
-    winc.style = CS_HREDRAW | CS_VREDRAW;
-    winc.lpfnWndProc = Some(win_proc);
-    winc.hIcon = LoadIconW(ptr::null_mut(), IDI_APPLICATION);
-    winc.hCursor = LoadCursorW(ptr::null_mut(), IDC_ARROW);
-    winc.hbrBackground = GetStockObject(WHITE_BRUSH as i32) as HBRUSH;
-    winc.lpszClassName = class_name.as_ptr();
-
-    RegisterClassW(&winc) > 0
-}
-
-unsafe extern "system" fn win_proc(
-    hwnd: HWND,
-    msg: UINT,
-    w_param: WPARAM,
-    l_param: LPARAM,
-) -> LRESULT {
-    match msg {
-        WM_DESTROY => PostQuitMessage(0),
-        _ => return DefWindowProcW(hwnd, msg, w_param, l_param),
-    };
-    0
 }
